requirements.txt
PyQt6
sounddevice
numpy
pynput
groq
python-dotenv
flask
pywin32
pyperclip
